breaking(on): switch to new decorator
