+++ src/decorators/on.js
@@ -9,56 +9,56 @@ declare var __DEV__: boolean
  * The decorator for registering event listener info to the method.
  * @param event The event name
  * @param at The selector
- * @param target The target prototype (decorator interface)
- * @param key The decorator target key (decorator interface)
+ * @param descriptor The method descriptor
  */
-export default (event: string, { at }: { at?: string } = {}) => (target: Object, key: string) => {
-  const Constructor = target.constructor
+export default (event: string, { at }: { at?: string } = {}) => (descriptor: Object) => {
+  const key = descriptor.key
 
-  check(!!event, `Empty event handler is given: constructor=${Constructor.name} key=${key}`)
+  descriptor.finisher = (constructor: Function) => {
+    check(!!event, `Empty event handler is given: constructor=${constructor.name} key=${key}`)
+    /**
+     * @param el The element
+     * @param coelem The coelement
+     * @param name The component name
+     */
+    constructor[KEY_EVENT_LISTENERS] = (constructor[KEY_EVENT_LISTENERS] || []).concat((el: HTMLElement, coelem: any, name: string) => {
+      const keyEventListeners = KEY_EVENT_LISTENERS + name
 
-  /**
-   * @param el The element
-   * @param coelem The coelement
-   * @param name The component name
-   */
-  Constructor[KEY_EVENT_LISTENERS] = (Constructor[KEY_EVENT_LISTENERS] || []).concat((el: HTMLElement, coelem: any, name: string) => {
-    const keyEventListeners = KEY_EVENT_LISTENERS + name
-
-    const listener = (e: Event): void => {
-      if (
-        !at ||
-        [].some.call(el.querySelectorAll(at), node => {
-          return node === e.target || node.contains(e.target)
-        })
-      ) {
-        if (__DEV__) {
-          debugMessage({
-            type: 'event',
-            module: '💊',
-            color: '#e0407b',
-            e,
-            el,
-            coelem
+      const listener = (e: Event): void => {
+        if (
+          !at ||
+          [].some.call(el.querySelectorAll(at), node => {
+            return node === e.target || node.contains(e.target)
           })
+        ) {
+          if (__DEV__) {
+            debugMessage({
+              type: 'event',
+              module: '💊',
+              color: '#e0407b',
+              e,
+              el,
+              coelem
+            })
+          }
+
+          coelem[key](e)
         }
-
-        coelem[key](e)
       }
-    }
 
-    /**
-     * Removes the event listener.
-     */
-    listener.remove = () => {
-      el.removeEventListener(event, listener)
-    }
+      /**
+       * Removes the event listener.
+       */
+      listener.remove = () => {
+        el.removeEventListener(event, listener)
+      }
 
-    /**
-     * Store event listeners to remove it later.
-     */
-    ;(el: any)[keyEventListeners] = ((el: any)[keyEventListeners] || []).concat(listener)
+      /**
+       * Store event listeners to remove it later.
+       */
+      ;(el: any)[keyEventListeners] = ((el: any)[keyEventListeners] || []).concat(listener)
 
-    el.addEventListener(event, listener)
-  })
+      el.addEventListener(event, listener)
+    })
+  }
 }
